Barrier
Working barrier using time. When the zombie touches the trigger next to the barrier a timer will count down when it reaches 0 it translate the barrier and the zombie can go out of it. If the player touches the trigger in front of the barrier entrence a timer will count down and when it reaches 0 the barrier is placed back. More barriers will be implemented when map is finished